Добавил генерацию комментариев в попап
