+++ src/mocks/comment.js
@@ -1,16 +1,19 @@
-import {getRandomArrayItem} from '../utils.js';
+import {
+  getRandomArrayItem,
+  getRandomDate,
+} from '../utils.js';
 
 
 const Comments = [
-  `Film name 1`,
-  `Film name 2`,
-  `Film name 3`,
+  `Comment 1`,
+  `Comment 2`,
+  `Comment 3`,
 ];
 
 const Emojis = [
-  `public\images\emoji\angry.png`,
-  `public\images\emoji\puke.png`,
-  `public\images\emoji\sleeping.png`,
+  `angry.png`,
+  `puke.png`,
+  `sleeping.png`,
 ];
 
 const Authors = [
@@ -18,14 +21,9 @@ const Authors = [
   `Tim Macoveev`,
 ];
 
+const COMMENT_DATE_MIN = new Date(`01.01.2010`);
+const COMMENT_DATE_MAX = new Date(`01.01.2020`);
 
-/**
- * Генерирует дату в формате `2019/12/31 23:59`
- * @return {string} Строка в формате `2019/12/31 23:59`
- */
-const generateRandomDate = () => {
-  return `2019/12/31 23:59`;
-};
 
 /**
  * Генерируем новый комментарий
@@ -34,13 +32,24 @@ const generateRandomDate = () => {
 const generateComment = () => {
   return {
     text: getRandomArrayItem(Comments),
-    emotion: `./public/images/emoji/` + getRandomArrayItem(Emojis),
+    emotion: getRandomArrayItem(Emojis),
     author: getRandomArrayItem(Authors),
-    date: generateRandomDate(),
+    date: getRandomDate(COMMENT_DATE_MIN, COMMENT_DATE_MAX),
   };
+};
+
+/**
+ * Генерируем массив комментариев
+ * @param {number} count - Число комментариев
+ * @return {object} - Массив комментариев
+ */
+const generateComments = (count) => {
+  return new Array(count)
+    .fill(``)
+    .map(generateComment);
 };
 
 
 export {
-  generateComment
+  generateComments
 };
