+++ src/mocks/film.js
@@ -2,7 +2,9 @@ import {
   getRandomArrayItem,
   getRandomIntegerNumber,
   castTimeFormat,
-  getRandomFloatNumber
+  getRandomFloatNumber,
+  getRandomArrayItems,
+  getRandomDate,
 } from '../utils.js';
 
 import {
@@ -13,7 +15,7 @@ import {
   FilmNames,
   Posters,
   Genre,
-  Descriptions
+  Descriptions,
 } from '../consts.js';
 
 
@@ -21,6 +23,50 @@ const MIN_RELEASE_YEAR = 1917;
 const MAX_RELEASE_YEAR = 2020;
 const MIN_COMMENTS = 1;
 const MAX_COMMENTS = 30;
+
+
+const AgeRatings = [
+  `0+`,
+  `6+`,
+  `12+`,
+  `18+`,
+];
+
+const Countries = [
+  `USA`,
+  `Russia`,
+  `New Zealand`,
+];
+
+const Directors = [
+  `Director1`,
+  `Director2`,
+  `Director3`,
+  `Director4`,
+  `Director5`,
+  `Director6`,
+];
+
+const Writers = [
+  `Screenwriter1`,
+  `Screenwriter2`,
+  `Screenwriter3`,
+  `Screenwriter4`,
+  `Screenwriter5`,
+  `Screenwriter6`,
+];
+
+const Actors = [
+  `Actor1`,
+  `Actor2`,
+  `Actor3`,
+  `Actor4`,
+  `Actor5`,
+  `Actor6`,
+];
+
+const RELEASE_DATE_MIN = new Date(`01.01.1900`);
+const RELEASE_DATE_MAX = new Date(`01.01.2020`);
 
 
 /**
@@ -37,9 +83,22 @@ const generateFilm = () => {
     genre: getRandomArrayItem(Genre),
     description: getRandomArrayItem(Descriptions),
     commentsNumber: getRandomIntegerNumber(MIN_COMMENTS, MAX_COMMENTS),
+
+    originalName: getRandomArrayItem(FilmNames),
+    director: getRandomArrayItem(Directors),
+    writers: getRandomArrayItems(Writers, 3),
+    actors: getRandomArrayItems(Actors, 4),
+    releaseDate: getRandomDate(RELEASE_DATE_MIN, RELEASE_DATE_MAX),
+    country: getRandomArrayItem(Countries),
+    ageRating: getRandomArrayItem(AgeRatings),
   };
 };
 
+/**
+ * Генерируем массив фильмов
+ * @param {number} count - Число фильмов
+ * @return {object} - Массив фильмов
+ */
 const generateFilms = (count) => {
   return new Array(count)
     .fill(``)
